+++ src/Contracts/Auth/Authorizer.php
@@ -11,6 +11,7 @@ declare(strict_types=1);
 
 namespace LaravelJsonApi\Contracts\Auth;
 
+use Illuminate\Auth\Access\Response;
 use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Auth\AuthenticationException;
 use Illuminate\Http\Request;
@@ -25,45 +26,45 @@ interface Authorizer
      *
      * @param Request|null $request
      * @param string $modelClass
-     * @return bool
+     * @return bool|Response
      */
-    public function index(?Request $request, string $modelClass): bool;
+    public function index(?Request $request, string $modelClass): bool|Response;
 
     /**
      * Authorize a JSON:API store operation.
      *
      * @param Request|null $request
      * @param string $modelClass
-     * @return bool
+     * @return bool|Response
      */
-    public function store(?Request $request, string $modelClass): bool;
+    public function store(?Request $request, string $modelClass): bool|Response;
 
     /**
      * Authorize a JSON:API show query.
      *
      * @param Request|null $request
      * @param object $model
-     * @return bool
+     * @return bool|Response
      */
-    public function show(?Request $request, object $model): bool;
+    public function show(?Request $request, object $model): bool|Response;
 
     /**
      * Authorize the update controller action.
      *
      * @param object $model
      * @param Request $request
-     * @return bool
+     * @return bool|Response
      */
-    public function update(Request $request, object $model): bool;
+    public function update(Request $request, object $model): bool|Response;
 
     /**
      * Authorize the destroy controller action.
      *
      * @param Request $request
      * @param object $model
-     * @return bool
+     * @return bool|Response
      */
-    public function destroy(Request $request, object $model): bool;
+    public function destroy(Request $request, object $model): bool|Response;
 
     /**
      * Authorize the show-related controller action.
@@ -71,9 +72,9 @@ interface Authorizer
      * @param Request $request
      * @param object $model
      * @param string $fieldName
-     * @return bool
+     * @return bool|Response
      */
-    public function showRelated(Request $request, object $model, string $fieldName): bool;
+    public function showRelated(Request $request, object $model, string $fieldName): bool|Response;
 
     /**
      * Authorize the show-relationship controller action.
@@ -81,9 +82,9 @@ interface Authorizer
      * @param Request $request
      * @param object $model
      * @param string $fieldName
-     * @return bool
+     * @return bool|Response
      */
-    public function showRelationship(Request $request, object $model, string $fieldName): bool;
+    public function showRelationship(Request $request, object $model, string $fieldName): bool|Response;
 
     /**
      * Authorize the update-relationship controller action.
@@ -91,9 +92,9 @@ interface Authorizer
      * @param Request $request
      * @param object $model
      * @param string $fieldName
-     * @return bool
+     * @return bool|Response
      */
-    public function updateRelationship(Request $request, object $model, string $fieldName): bool;
+    public function updateRelationship(Request $request, object $model, string $fieldName): bool|Response;
 
     /**
      * Authorize the attach-relationship controller action.
@@ -101,9 +102,9 @@ interface Authorizer
      * @param Request $request
      * @param object $model
      * @param string $fieldName
-     * @return bool
+     * @return bool|Response
      */
-    public function attachRelationship(Request $request, object $model, string $fieldName): bool;
+    public function attachRelationship(Request $request, object $model, string $fieldName): bool|Response;
 
     /**
      * Authorize the detach-relationship controller action.
@@ -111,9 +112,9 @@ interface Authorizer
      * @param Request $request
      * @param object $model
      * @param string $fieldName
-     * @return bool
+     * @return bool|Response
      */
-    public function detachRelationship(Request $request, object $model, string $fieldName): bool;
+    public function detachRelationship(Request $request, object $model, string $fieldName): bool|Response;
 
     /**
      * Get JSON:API errors describing the failure, or throw an appropriate exception.