+++ src/Core/Exceptions/JsonApiException.php
@@ -38,7 +38,7 @@ class JsonApiException extends Exception implements HttpExceptionInterface, Resp
      * @param Throwable|null $previous
      * @return static
      */
-    public static function make($errors, Throwable $previous = null): self
+    public static function make($errors, ?Throwable $previous = null): self
     {
         return new self($errors, $previous);
     }
@@ -50,7 +50,7 @@ class JsonApiException extends Exception implements HttpExceptionInterface, Resp
      * @param Throwable|null $previous
      * @return static
      */
-    public static function error($error, Throwable $previous = null): self
+    public static function error($error, ?Throwable $previous = null): self
     {
         return new self(Error::cast($error), $previous);
     }
@@ -62,7 +62,7 @@ class JsonApiException extends Exception implements HttpExceptionInterface, Resp
      * @param Throwable|null $previous
      * @param array $headers
      */
-    public function __construct($errors, Throwable $previous = null, array $headers = [])
+    public function __construct($errors, ?Throwable $previous = null, array $headers = [])
     {
         parent::__construct('JSON:API error', 0, $previous);
         $this->errors = ErrorList::cast($errors);